packages/cli: most of watch-deps into lib and re-use from there instead
