+++ packages/cli/src/lib/watchDeps/compiler.ts
