+++ packages/cli/src/lib/watchDeps/index.ts
@@ -14,20 +14,4 @@
  * limitations under the License.
  */
 
-import { spawn } from 'child_process';
-import { createLogPipe } from 'lib/logging';
-
-export function startChild(args: string[]) {
-  const [command, ...commandArgs] = args;
-  const child = spawn(command, commandArgs, {
-    env: { FORCE_COLOR: 'true', ...process.env },
-    stdio: ['inherit', 'pipe', 'pipe'],
-    shell: true,
-  });
-
-  // We need to avoid clearing the terminal, or the build feedback of dependencies will be lost
-  const logPipe = createLogPipe();
-  child.stdout.on('data', logPipe(process.stdout));
-  child.stderr.on('data', logPipe(process.stderr));
-  return child;
-}
+export * from './watchDeps';
