+++ packages/cli/src/lib/watchDeps/packages.ts
