+++ packages/cli/src/lib/watchDeps/watcher.ts
